🧹 Tidy: Contraction機能のリファクタリング
- Contraction関連コンポーネントを `components/contraction/` に集約
- `ContractionWaveGraph` のロジックを `useContractionWaveData` フックに抽出
- `ContractionHistory` のロジックを `useContractionActions` フックに抽出
- コンポーネントの責務分離と可読性向上を実現
