+++ frontend/app/(dashboard)/contraction/page.tsx
@@ -2,10 +2,10 @@
 
 import { useContractions } from "@/hooks/useData"
 import { useRecordPage } from "@/hooks/useRecordPage"
-import ContractionTimer from "@/components/ContractionTimer"
-import ContractionStats from "@/components/ContractionStats"
-import ContractionHistory from "@/components/ContractionHistory"
-import ContractionWaveGraph from "@/components/ContractionWaveGraph"
+import ContractionTimer from "@/components/contraction/ContractionTimer"
+import ContractionStats from "@/components/contraction/ContractionStats"
+import ContractionHistory from "@/components/contraction/ContractionHistory"
+import ContractionWaveGraph from "@/components/contraction/ContractionWaveGraph"
 import type { ContractionRecord } from "@/types/contraction"
 import { TipsCard } from "@/components/ui/tips-card"
 import { contractionTips } from "@/lib/tips-data"
