+++ frontend/components/contraction/ContractionStats.tsx
