+++ frontend/components/contraction/ContractionTimer.tsx
